fix: Migration — platforms table, DemoSeeder role
- Add description, icon, color, config, is_published, sort_order to platforms
- Update Platform model and seeder to use name instead of slug
- Fix DemoSeeder: change 'staff' role to 'member' (staff doesn't exist)
- Fix approved_by and client_id to unsignedBigInteger for FK compatibility

migrate:fresh --seed now works correctly

+++ database/migrations/2026_09_16_000000_add_approval_to_social_posts.php
@@ -11,10 +11,10 @@ return new class extends Migration
     {
         Schema::table('social_posts', function (Blueprint $table) {
             $table->string('approval_status')->nullable()->after('status');
-            $table->integer('approved_by')->nullable()->after('approval_status');
+            $table->unsignedBigInteger('approved_by')->nullable()->after('approval_status');
             $table->timestamp('approved_at')->nullable()->after('approved_by');
             $table->text('approval_notes')->nullable()->after('approved_at');
-            $table->integer('client_id')->nullable()->after('approval_notes');
+            $table->unsignedBigInteger('client_id')->nullable()->after('approval_notes');
         });
     }
